+++ rrule/src/lib.rs
@@ -40,6 +40,7 @@
 //! use rrule::RRuleSet;
 //!
 //! let rrule: RRuleSet = "DTSTART:20120201T093000Z\nRRULE:FREQ=DAILY;COUNT=3".parse().unwrap();
+//! let (events, _) = rrule.all(100);
 //!
 //! // All dates
 //! assert_eq!(
@@ -48,7 +49,7 @@
 //!         DateTime::parse_from_rfc3339("2012-02-02T09:30:00+00:00").unwrap(),
 //!         DateTime::parse_from_rfc3339("2012-02-03T09:30:00+00:00").unwrap(),
 //!     ],
-//!     rrule.all(100).unwrap()
+//!     events
 //! );
 //! ```
 //! Find all events that are within a given range.
@@ -64,13 +65,14 @@
 //! let before = UTC.ymd(2012, 4, 1).and_hms(9, 0, 0);
 //!
 //! let rrule = rrule.after(after).before(before);
+//! let (events, _) = rrule.all(100);
 //!
 //! assert_eq!(
 //!     vec![
 //!         DateTime::parse_from_rfc3339("2012-02-02T09:30:00+00:00").unwrap(),
 //!         DateTime::parse_from_rfc3339("2012-02-03T09:30:00+00:00").unwrap(),
 //!     ],
-//!     rrule.all(100).unwrap()
+//!     events
 //! );
 //! ```
 //!
@@ -92,6 +94,6 @@ mod validator;
 pub use crate::core::{Frequency, NWeekday, RRule, RRuleSet};
 pub use crate::core::{Unvalidated, Validated};
 pub use chrono::Weekday;
-pub use error::{RRuleError, WithError};
+pub use error::RRuleError;
 pub(crate) use iter::RRuleIter;
 pub use iter::RRuleSetIter;