+++ rrule/src/validator/error.rs
@@ -42,6 +42,10 @@ pub enum ValidationError {
             See 'validator limits' in docs for more info."
     )]
     StartYearOutOfRange(i32),
+    #[error(
+        "Unable to generate a timeset for the RRULE. Please specify a BYHOUR, BYMINUTE or BYSECOND"
+    )]
+    UnableToGenerateTimeset,
     #[cfg(feature = "by-easter")]
     #[error("`BYEASTER` can only be used when `BYHOUR`, `BYMINUTE` and `BYSECOND` are set.")]
     InvalidByRuleWithByEaster,